refactor: 패키지 분리
vendingMachine을 어플리케이션 전체에서 공유함에 따라 서비스간의
의존관계가 사라져 패키지를 도메인별로 분리
