+++ src/main/java/vendingmachine/purchase/PurchaseController.java
@@ -1,4 +1,6 @@
-package vendingmachine;
+package vendingmachine.purchase;
+
+import vendingmachine.VendingMachineConsole;
 
 public class PurchaseController {
     private final PurchaseService purchaseService;
